requirements.txt
streamlit
faster-whisper
pyannote.audio
torch
torchaudio
pydub
requests
